jeans-api/requirements.txt
Flask==2.0.0
Werkzeug==2.0.0
pymongo==3.12.1
python-dotenv==0.19.0
flask-pymongo==2.3.0
dnspython==2.1.0  